requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-html==4.1.1
pytest-json-report==1.5.0
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0
elasticsearch==8.10.0
